model/requirements.txt
influxdb-client
faker
google-cloud-bigquery
flask-apscheduler
tensorflow